refactor(generation): move shared runtimes into services
